Move test library files to lib
