+++ e2e/app/playwright.config.js
@@ -1,5 +1,5 @@
 import baseConfig from '../playwright.config';
-import { deepMerge } from '../util';
+import { deepMerge } from '../lib/util';
 import { defineConfig } from '@playwright/test';
 
 export default defineConfig(deepMerge(
